+++ backend/src/test/java/middle_point_search/backend/domains/member/LoginMemberIntegrationTest.java
@@ -11,8 +11,6 @@ import org.springframework.http.MediaType;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.web.servlet.ResultActions;
 
-import com.fasterxml.jackson.databind.JsonNode;
-
 import middle_point_search.backend.common.BaseIntegrationTest;
 import middle_point_search.backend.domains.member.domain.Member;
 import middle_point_search.backend.domains.member.domain.Role;
@@ -83,32 +81,5 @@ public class LoginMemberIntegrationTest extends BaseIntegrationTest {
 			.andExpect(status().isOk())
 			.andExpect(jsonPath("$.data.accessToken").doesNotExist())
 			.andExpect(jsonPath("$.data.refreshToken").doesNotExist());
-	}
-
-	@Test
-	@DisplayName("엑세스 토큰 재발급에 성공한다.")
-	public void 엑세스토큰재발급성공() throws Exception {
-		// 로그인
-		LoginMemberRequest loginMemberRequest = new LoginMemberRequest(email, pw);
-		ResultActions resultActions = mockMvc.perform(post("/api/members/login")
-			.contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-			.param("email", loginMemberRequest.email())
-			.param("pw", loginMemberRequest.pw())
-			.accept(MediaType.APPLICATION_JSON));
-
-		String responseContent = resultActions.andReturn().getResponse().getContentAsString();
-		JsonNode jsonNode = objectMapper.readTree(responseContent);
-		String accessToken = jsonNode.get("data").get("accessToken").asText();
-		String refreshToken = jsonNode.get("data").get("refreshToken").asText();
-
-		// when
-		ResultActions resultActions1 = mockMvc.perform(post("/api/members/refresh-token")
-			.header("Authorization-refresh", "Bearer " + refreshToken));
-
-		// then
-		resultActions1
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.data.accessToken").exists())
-			.andExpect(jsonPath("$.data.refreshToken").exists());
 	}
 }
